Verified

+++ project.py
@@ -275,10 +275,6 @@ def newTourism():
         return redirect(url_for('places'))
     else:
         return render_template('newTourism.html')
-    if Tourism.user_id != login_session['user_id']:
-        return "<script>function myFunction(){alert('You are not authorized to\
-        create this touristplace. please create your own touristplace in order\
-        to create.');}</script><body onload='myFunction()'>"
     # return "This page will be for making a newtourists"
     # Edit a tourist place
 
@@ -289,6 +285,10 @@ def editTourism(tourist_id):
         return redirect('/login')
     editedTourism = \
         session.query(Tourism).filter_by(id=tourist_id).one_or_none()
+    if Tourism.user_id != login_session['user_id']:
+        return "<script>function myFunction(){alert('You are not authorized to\
+        edit this touristplace. please create your own touristplace in order\
+        to edit.');}</script><body onload='myFunction()'>"
     if request.method == 'POST':
         if request.form['name']:
             editedTourism.name = request.form['name']
@@ -296,10 +296,6 @@ def editTourism(tourist_id):
     else:
         return render_template('editTourism.html',
                                tourist=editedTourism)
-    if Tourism.user_id != login_session['user_id']:
-        return "<script>function myFunction(){alert('You are not authorized to\
-        edit this touristplace. please create your own touristplace in order\
-        to edit.');}</script><body onload='myFunction()'>"
     # return 'This page will be for editing tourist %s' %tourists_id
     # Delete a tourist place
 
@@ -310,6 +306,10 @@ def deleteTourism(tourist_id):
         return redirect('/login')
     touristToDelete = \
         session.query(Tourism).filter_by(id=tourist_id).one_or_none()
+    if Tourism.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('you are not authorized to\
+         delete this tourist.please create your own tourist to delete');}\
+         </script><body onLoad = 'myFunction()'>"
     if request.method == 'POST':
         session.delete(touristToDelete)
         session.commit()
@@ -317,10 +317,6 @@ def deleteTourism(tourist_id):
     else:
         return render_template('deleteTourism.html',
                                tourist=touristToDelete)
-    if Tourism.user_id != login_session['user_id']:
-        return "<script>function myFunction() {alert('you are not authorized to\
-         delete this tourist.please create your own tourist to delete');}\
-         </script><body onLoad = 'myFunction()'>"
     # return 'This page will be for deleting tourist %s' % tourists_id
     ''' Show a tourist spot'''
 
@@ -354,12 +350,8 @@ def newDestination(tourist_id):
         return redirect(url_for('showplace', tourist_id=tourist_id))
     else:
         return render_template('newspot.html', tourist_id=tourist_id)
-
     return render_template('newDestination.html', tourist=tourist)
-    if login_session['user_id'] != Destination.user_id:
-        return "<script>function myFunction() {alert('You are not authorized to\
-          create spots to this tourist.Please create your own tourist in\
-          order to create spots.');}</script><body onload='myFunction()'>"
+
     # return 'This page is for making a new spot for tourist places %s'
     # %tourist_id
     # Edit a spot
@@ -371,6 +363,10 @@ def editDestination(tourist_id, area_id):
     if 'username' not in login_session:
         return redirect('/login')
     editedlook = session.query(Destination).filter_by(id=area_id).one_or_none()
+    if login_session['user_id'] != Destination.user_id:
+        return "<script>function myFunction() {alert('You are not authorized to\
+          edit spots to this tourist.Please create your own tourist in\
+          order to edit spots.');}</script><body onload='myFunction()'>"
     if request.method == 'POST':
         if request.form['name']:
             editedlook.name = request.form['name']
@@ -384,13 +380,8 @@ def editDestination(tourist_id, area_id):
         session.commit()
         return redirect(url_for('showplace', tourist_id=tourist_id))
     else:
-
         return render_template('editspot.html', tourist_id=tourist_id,
                                area_id=area_id, look=editedlook)
-    if login_session['user_id'] != Destination.user_id:
-        return "<script>function myFunction() {alert('You are not authorized to\
-          edit spots to this tourist.Please create your own tourist in\
-          order to edit spots.');}</script><body onload='myFunction()'>"
         # return 'This page is for editing spot %s' % area_id
         # Delete a spot
 
@@ -402,7 +393,10 @@ def deleteDestination(tourist_id, area_id):
         return redirect('/login')
     lookToDelete = \
         session.query(Destination).filter_by(id=area_id).one_or_none()
-
+    if login_session['user_id'] != Destination.user_id:
+        return "<script>function myFunction() {alert ('you are not authorized to\
+         delete spots to this tourist.please create your own tourist\
+         in order to delete spots');}</script><body onload='myFunction()'>"
     if request.method == 'POST':
         session.delete(lookToDelete)
         session.commit()
@@ -410,11 +404,7 @@ def deleteDestination(tourist_id, area_id):
     else:
         return render_template('deleteDestination.html',
                                look=lookToDelete)
-    if login_session['user_id'] != Destination.user_id:
-        return "<script>function myFunction() {alert ('you are not authorized to\
-         delete spots to this tourist.please create your own tourist\
-         in order to delete spots');}</script><body onload='myFunction()'>"
-    # return "This page is for deleting spot %s" %area_id
+     # return "This page is for deleting spot %s" %area_id
 
 if __name__ == '__main__':
     app.secret_key = 'super_secret_key'
